Pass --set-samba-secret to msktutil whenever it makes sense.
Today we found (in INC0207235) that a client was unable to connect to
Samba. Typing "net ads join" manually solved the problem, but only for
a short period of time (roughly 24h).

Even though we haven't been able to reproduce that failure
mode (neither with a Mac, with or without a Kerberos ticket visible in
"klist"; nor with a Windows 10 VM joined into the domain), we have a
theory on the cause: we suspect that Samba renews the machine trust
password using the old password as the means of authentication, while
msktutil does so using the Kerberos keytab... and forgets about the
new machine trust password, which it doesn't need. Hence, we surmise
the following sequence of events, causing the outage:

* the daily crontab was running msktutil *without* the
  --set-samba-secret flag, changing the trust account password and
  promptly forgetting the new one
* sometime later, Samba attempts to renew the trust account password,
  and fails (since it is no longer current)
* at this point in time, Samba gives up on a certain category of logon
  methods (we assume), causing the symptoms of INC0207235.

Changed in this commit:
* New fact $::has_net_changesecretpw which informs the default value
  of $::epfl_sso::private::ad::set_samba_secret
* When true, pass --set-samba-secret to msktutil

TESTED: "net ads join" works OK after the new crontab runs; didn't
with the old one (with the caveat that one needs to replace
--auto-update with -u in both tests, to circumvent the check for
password age built into msktutil)